refactor: update schema imports and enhance login schema validation
