+++ src/schemas/index.ts
@@ -1,2 +1,3 @@
 export { EmailSchema } from "./base-schema";
-export { default as LoginSchema } from "./auth/login-schema";
+export { LoginSchema, type LoginSchemaType } from "./auth/login-schema";
+export { SignUpSchema, type SignUpSchemaType } from "./auth/sign-up-schema";
